[*] move core dns logic under `dns/`
