+++ dns/parser.go
@@ -1,4 +1,4 @@
-package main
+package dns
 
 import (
 	"bytes"
